Add methods to controllers

+++ src/main/kotlin/com/history/tree/model/RelationshipType.kt
@@ -4,18 +4,18 @@ import org.springframework.core.convert.converter.Converter
 import org.springframework.data.convert.ReadingConverter
 import org.springframework.data.convert.WritingConverter
 
-enum class RelationshipType(val code: Int) {
+enum class RelationshipType(val code: Short) {
     MARRIAGE(1), EX_MARRIAGE(2), PARENT(3)
 }
 
 @WritingConverter
-class RelationshipTypeToIntConverter : Converter<RelationshipType, Int> {
-    override fun convert(source: RelationshipType): Int = source.code
+class RelationshipTypeToIntConverter : Converter<RelationshipType, Short> {
+    override fun convert(source: RelationshipType): Short = source.code
 }
 
 @ReadingConverter
-class IntToRelationshipTypeConverter : Converter<Int, RelationshipType> {
-    override fun convert(source: Int): RelationshipType {
+class IntToRelationshipTypeConverter : Converter<Short, RelationshipType> {
+    override fun convert(source: Short): RelationshipType {
         return RelationshipType.values().first { p -> p.code == source }
     }
 }
